Changes the cursor data structure from a list to a map
With the changes in this commit cursors will become maps where keys
represent columns and values represent the values of the cursor exactly
as they were in the previous list cursors. Example:

```
record = %{id: "123", date: "2020-08-01"}
cursor = Paginator.cursor_for_record(record, [:id, :date])
```

Before the changes in this commit cursor in the example above will be
`["123", "2020-08-01"]`. With the changes in this commit the cursor
value will be `%{id: "123", date: "2020-08-01"}`.

This change will make it possible to validate that the cursor columns
match the cursor fields used.

This change is backward compatible: users can use old list cursors like
they normally would, and will get new cursors in the page result.