Adds validation to guarantee that the cursor(s) match the cursor fields
This validation prevents users to pass in cursors that don't match the
cursor fields used, which in other words means that it prevents
Paginator from generating wrong queries. Example:

```
record = %{id: "123", date: "2020-08-01"}
cursor = Paginator.cursor_for_record(record, [:id, :date])
page = Paginator.paginate(cursor_fields: [:date], after: cursor)
```

The example above will produce an invlid query because the columns in
the cursor don't match the columns in the cursor_fields. In practice
Paginator will generate a query with a filter on the date based on the
value of the id.

With the changes in this commit the code above will raise an error
preventing Paginator from generating the wrong query.