Fixes bug that generates queries with cursor conditions out of order
In https://github.com/duffelhq/paginator/pull/68 we changed the cursor
data structure from a list to a map. Because elements in a map don't
have order among them, this change introduced a bug that causes
paginator to generate pagination queries with the cursor conditions
following the order in which they are stored in a map (which is random).

This commit fixes this by enforcing the cursor_fields order on the
cursor values when generating the pagination queries.