feat: add support for all asc/desc sorting order combinations
This patch adds support for all supported sorting orders when the
field is nullable. It is now possible to specify the following extra
sort orders in the cursor_fields:

  * asc_nulls_last, asc_nulls_first
  * desc_nulls_last, desc_nulls_first

As the number of different combinations is large, we refactored the
code that builds the dynamic filter expressions to make it a bit more
manageable. Each sorting order is now a separate module.

There is at least one problem with the current implementation
though. The `:asc/:desc` default order, w.r.t. to null values may vary
from vendor to vendor. For instance, we are following Postgres
defaults. It means that `:asc = :asc_nulls_last` and `:desc =
:desc_nulls_first`. However, if you are using Mysql instead, you get
`:asc = :asc_nulls_first` and `:desc = :desc_nulls_last`. Somehow we
need to factor in the adapter. However, I was not sure the best way to
achieve that, I opted for lefting it out for now and instead asks for
directions on that regard.